bypass scheduler working!
