+++ src/object/notification.c
@@ -32,34 +32,6 @@ static inline void ntfn_ptr_set_queue(notification_t *ntfnPtr, tcb_queue_t ntfn_
     notification_ptr_set_ntfnQueue_tail(ntfnPtr, (word_t)ntfn_queue.end);
 }
 
-static inline void ntfn_set_active(notification_t *ntfnPtr, word_t badge)
-{
-    notification_ptr_set_state(ntfnPtr, NtfnState_Active);
-    notification_ptr_set_ntfnMsgIdentifier(ntfnPtr, badge);
-}
-
-#ifdef CONFIG_KERNEL_MCS
-static inline void maybeDonateSchedContext(tcb_t *tcb, notification_t *ntfnPtr)
-{
-    if (tcb->tcbSchedContext == NULL) {
-        sched_context_t *sc = SC_PTR(notification_ptr_get_ntfnSchedContext(ntfnPtr));
-        if (sc != NULL && sc->scTcb == NULL) {
-            schedContext_donate(sc, tcb);
-            if (sc != NODE_STATE(ksCurSC)) {
-                /* refill_unblock_check should not be called on the
-                 * current SC as it is already running. The current SC
-                 * may have been bound to a notificaiton object if the
-                 * current thread was deleted in a long-running deletion
-                 * that became preempted. */
-                refill_unblock_check(sc);
-            }
-            schedContext_resume(sc);
-        }
-    }
-}
-
-#endif
-
 #ifdef CONFIG_KERNEL_MCS
 #define MCS_DO_IF_SC(tcb, ntfnPtr, _block) \
     maybeDonateSchedContext(tcb, ntfnPtr); \
